Fix several warnings from ex_doc
