+++ lib/open_api_spex/plug/app_env_cache.ex
@@ -1,0 +1,18 @@
+defmodule OpenApiSpex.Plug.AppEnvCache do
+  @behaviour OpenApiSpex.Plug.Cache
+
+  @impl true
+  def get(spec_module) do
+    Application.get_env(:open_api_spex, spec_module)
+  end
+
+  @impl true
+  def put(spec_module, spec) do
+    :ok = Application.put_env(:open_api_spex, spec_module, spec)
+  end
+
+  @impl true
+  def erase(spec_module) do
+    Application.delete_env(:open_api_spex, spec_module)
+  end
+end
