+++ lib/open_api_spex/plug/cache.ex
@@ -53,63 +53,3 @@ defmodule OpenApiSpex.Plug.Cache do
     adapter.erase()
   end
 end
-
-defmodule OpenApiSpex.Plug.AppEnvCache do
-  @moduledoc false
-  @behaviour OpenApiSpex.Plug.Cache
-
-  @impl true
-  def get(spec_module) do
-    Application.get_env(:open_api_spex, spec_module)
-  end
-
-  @impl true
-  def put(spec_module, spec) do
-    :ok = Application.put_env(:open_api_spex, spec_module, spec)
-  end
-
-  @impl true
-  def erase(spec_module) do
-    Application.delete_env(:open_api_spex, spec_module)
-  end
-end
-
-if function_exported?(:persistent_term, :info, 0) do
-  defmodule OpenApiSpex.Plug.PersistentTermCache do
-    @moduledoc false
-    @behaviour OpenApiSpex.Plug.Cache
-
-    @impl true
-    def get(spec_module) do
-      :persistent_term.get(spec_module)
-    rescue
-      ArgumentError ->
-        nil
-    end
-
-    @impl true
-    def put(spec_module, spec) do
-      :ok = :persistent_term.put(spec_module, spec)
-    end
-
-    @impl true
-    def erase(spec_module) do
-      :persistent_term.erase(spec_module)
-      :ok
-    end
-  end
-end
-
-defmodule OpenApiSpex.Plug.NoneCache do
-  @moduledoc false
-  @behaviour OpenApiSpex.Plug.Cache
-
-  @impl true
-  def get(_spec_module), do: nil
-
-  @impl true
-  def put(_spec_module, _spec), do: :ok
-
-  @impl true
-  def erase(_spec_module), do: :ok
-end
