+++ lib/open_api_spex/plug/none_cache.ex
@@ -1,0 +1,12 @@
+defmodule OpenApiSpex.Plug.NoneCache do
+  @behaviour OpenApiSpex.Plug.Cache
+
+  @impl true
+  def get(_spec_module), do: nil
+
+  @impl true
+  def put(_spec_module, _spec), do: :ok
+
+  @impl true
+  def erase(_spec_module), do: :ok
+end
